Added tests for the `memlet_to_map()` function.
In fact I just relocated some of the `CopyToMap` tests.
